+++ cli/cmd/uninstall/uninstall.go
@@ -16,7 +16,14 @@ import (
 	"golang.org/x/text/language"
 	"helm.sh/helm/v3/pkg/action"
 	helmCLI "helm.sh/helm/v3/pkg/cli"
+	apiextv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	apiext "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 )
 
@@ -34,7 +41,7 @@ const (
 	defaultWipeData = false
 
 	flagTimeout    = "timeout"
-	defaultTimeout = "10m"
+	defaultTimeout = 10 * time.Minute
 
 	flagContext    = "context"
 	flagKubeconfig = "kubeconfig"
@@ -45,7 +52,10 @@ type Command struct {
 
 	helmActionsRunner helm.HelmActionsRunner
 
-	kubernetes kubernetes.Interface
+	// Configuration for interacting with Kubernetes.
+	k8sClient        kubernetes.Interface
+	dynamicK8sClient dynamic.Interface
+	apiextK8sClient  apiext.Interface
 
 	set *flag.Sets
 
@@ -53,8 +63,7 @@ type Command struct {
 	flagReleaseName string
 	flagAutoApprove bool
 	flagWipeData    bool
-	flagTimeout     string
-	timeoutDuration time.Duration
+	flagTimeout     time.Duration
 
 	flagKubeConfig  string
 	flagKubeContext string
@@ -90,7 +99,7 @@ func (c *Command) init() {
 		Default: defaultAnyReleaseName,
 		Usage:   "Name of the installation. This can be used to uninstall and/or delete the resources of a specific Helm release.",
 	})
-	f.StringVar(&flag.StringVar{
+	f.DurationVar(&flag.DurationVar{
 		Name:    flagTimeout,
 		Target:  &c.flagTimeout,
 		Default: defaultTimeout,
@@ -144,12 +153,6 @@ func (c *Command) Run(args []string) int {
 		c.UI.Output("Can't set -wipe-data alone. Omit this flag to interactively uninstall, or use it with -auto-approve to wipe all data during the uninstall.", terminal.WithErrorStyle())
 		return 1
 	}
-	duration, err := time.ParseDuration(c.flagTimeout)
-	if err != nil {
-		c.UI.Output("unable to parse -%s: %s", flagTimeout, err, terminal.WithErrorStyle())
-		return 1
-	}
-	c.timeoutDuration = duration
 
 	// helmCLI.New() will create a settings object which is used by the Helm Go SDK calls.
 	settings := helmCLI.New()
@@ -160,20 +163,8 @@ func (c *Command) Run(args []string) int {
 		settings.KubeContext = c.flagKubeContext
 	}
 
-	// Set up the kubernetes client to use for non Helm SDK calls to the Kubernetes API
-	// The Helm SDK will use settings.RESTClientGetter for its calls as well, so this will
-	// use a consistent method to target the right cluster for both Helm SDK and non Helm SDK calls.
-	if c.kubernetes == nil {
-		restConfig, err := settings.RESTClientGetter().ToRESTConfig()
-		if err != nil {
-			c.UI.Output("retrieving Kubernetes auth: %v", err, terminal.WithErrorStyle())
-			return 1
-		}
-		c.kubernetes, err = kubernetes.NewForConfig(restConfig)
-		if err != nil {
-			c.UI.Output("initializing Kubernetes client: %v", err, terminal.WithErrorStyle())
-			return 1
-		}
+	if err := c.initKubernetes(settings); err != nil {
+		c.UI.Output("Could not initialize Kubernetes client: %v", err, terminal.WithErrorStyle())
 	}
 
 	// Setup logger to stream Helm library logs.
@@ -183,7 +174,7 @@ func (c *Command) Run(args []string) int {
 	}
 
 	actionConfig := new(action.Configuration)
-	actionConfig, err = helm.InitActionConfig(actionConfig, c.flagNamespace, settings, uiLogger)
+	actionConfig, err := helm.InitActionConfig(actionConfig, c.flagNamespace, settings, uiLogger)
 	if err != nil {
 		c.UI.Output(err.Error(), terminal.WithErrorStyle())
 		return 1
@@ -319,6 +310,36 @@ func (c *Command) Run(args []string) int {
 	return 0
 }
 
+// initKubernetes sets up the kubernetes clients to use for non Helm SDK calls to the Kubernetes API.
+// The Helm SDK will use settings.RESTClientGetter for its calls as well, so this will
+// use a consistent method to target the right cluster for both Helm SDK and non Helm SDK calls.
+func (c *Command) initKubernetes(settings *helmCLI.EnvSettings) error {
+	restConfig, err := settings.RESTClientGetter().ToRESTConfig()
+	if err != nil {
+		return err
+	}
+
+	if c.k8sClient == nil {
+		if c.k8sClient, err = kubernetes.NewForConfig(restConfig); err != nil {
+			return err
+		}
+	}
+
+	if c.dynamicK8sClient == nil {
+		if c.dynamicK8sClient, err = dynamic.NewForConfig(restConfig); err != nil {
+			return err
+		}
+	}
+
+	if c.apiextK8sClient == nil {
+		if c.apiextK8sClient, err = apiext.NewForConfig(restConfig); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
 func (c *Command) uninstallHelmRelease(releaseName, namespace, releaseType string, settings *helmCLI.EnvSettings,
 	uiLogger action.DebugLog, actionConfig *action.Configuration) error {
 	c.UI.Output(fmt.Sprintf("Existing %s installation found.", releaseType), terminal.WithSuccessStyle())
@@ -343,13 +364,21 @@ func (c *Command) uninstallHelmRelease(releaseName, namespace, releaseType strin
 		}
 	}
 
+	// Delete any custom resources managed by Consul. If they cannot be deleted,
+	// patch the finalizers to be empty on each one.
+	if releaseType == common.ReleaseTypeConsul {
+		if err := c.removeCustomResources(uiLogger); err != nil {
+			c.UI.Output("Error removing custom resources: %v", err.Error(), terminal.WithErrorStyle())
+		}
+	}
+
 	actionConfig, err := helm.InitActionConfig(actionConfig, namespace, settings, uiLogger)
 	if err != nil {
 		return err
 	}
 
 	uninstall := action.NewUninstall(actionConfig)
-	uninstall.Timeout = c.timeoutDuration
+	uninstall.Timeout = c.flagTimeout
 
 	res, err := c.helmActionsRunner.Uninstall(uninstall, releaseName)
 	if err != nil {
@@ -359,7 +388,167 @@ func (c *Command) uninstallHelmRelease(releaseName, namespace, releaseType strin
 		c.UI.Output("Uninstall result: %s", res.Info, terminal.WithInfoStyle())
 		return nil
 	}
+
 	c.UI.Output(fmt.Sprintf("Successfully uninstalled %s Helm release.", releaseType), terminal.WithSuccessStyle())
+	return nil
+}
+
+// removeCustomResources fetches a list of custom resource defintions managed
+// by Consul and attempts to delete every custom resource for each definition.
+// If the resources cannot be deleted directly, the finalizers on each resource
+// are patched to be an empty list, freeing them to be deleted by Kubernetes.
+func (c *Command) removeCustomResources(uiLogger action.DebugLog) error {
+	uiLogger("Deleting custom resources managed by Consul")
+
+	crds, err := c.fetchCustomResourceDefinitions()
+	if err != nil {
+		return fmt.Errorf("unable to fetch Custom Resource Definitions for Consul deployment: %v", err)
+	}
+	kindToResource := mapCRKindToResourceName(crds)
+
+	crs, err := c.fetchCustomResources(crds)
+	if err != nil {
+		return err
+	}
+
+	if err = c.deleteCustomResources(crs, kindToResource, uiLogger); err != nil {
+		return err
+	}
+
+	err = backoff.Retry(func() error {
+		crs, err := c.fetchCustomResources(crds)
+		if err != nil {
+			return err
+		}
+		if len(crs) != 0 {
+			return common.NewDanglingResourceError(fmt.Sprintf("%d custom resources remain after deletion request", len(crs)))
+		}
+		return nil
+	}, backoff.WithMaxRetries(backoff.NewConstantBackOff(time.Second), 5))
+	if !common.IsDanglingResourceError(err) {
+		return err
+	}
+
+	// Custom resources could not be deleted directly, attempt to patch their finalizers to an empty array.
+	uiLogger("Patching finalizers on custom resources managed by Consul")
+
+	crs, err = c.fetchCustomResources(crds)
+	if err != nil {
+		return err
+	}
+
+	if err = c.patchCustomResources(crs, kindToResource, uiLogger); err != nil {
+		return err
+	}
+
+	err = backoff.Retry(func() error {
+		crs, err := c.fetchCustomResources(crds)
+		if err != nil {
+			return err
+		}
+		if len(crs) != 0 {
+			return common.NewDanglingResourceError(fmt.Sprintf("%d custom resources remain after request to patch finalizers", len(crs)))
+		}
+		return nil
+	}, backoff.WithMaxRetries(backoff.NewConstantBackOff(time.Second), 5))
+	if err != nil {
+		return fmt.Errorf("unable to remove all custom resources managed by Consul. %d custom resources remain and will need to be removed manually. %v", len(crs), err)
+	}
+
+	return nil
+}
+
+// fetchCustomResourceDefinitions fetches all Custom Resource Definitions managed by Consul.
+func (c *Command) fetchCustomResourceDefinitions() (*apiextv1.CustomResourceDefinitionList, error) {
+	return c.apiextK8sClient.ApiextensionsV1().CustomResourceDefinitions().List(c.Ctx, metav1.ListOptions{
+		LabelSelector: "app=consul",
+	})
+}
+
+// fetchCustomResources gets a list of all custom resources deployed in the
+// cluster that are managed by Consul.
+func (c *Command) fetchCustomResources(crds *apiextv1.CustomResourceDefinitionList) ([]unstructured.Unstructured, error) {
+	crs := make([]unstructured.Unstructured, 0)
+	for _, crd := range crds.Items {
+		for _, version := range crd.Spec.Versions {
+			target := schema.GroupVersionResource{
+				Group:    crd.Spec.Group,
+				Version:  version.Name,
+				Resource: crd.Spec.Names.Plural,
+			}
+
+			crList, err := c.dynamicK8sClient.Resource(target).List(c.Ctx, metav1.ListOptions{})
+			if err != nil {
+				return nil, err
+			}
+			if crList != nil {
+				crs = append(crs, crList.Items...)
+			}
+		}
+	}
+
+	return crs, nil
+}
+
+// deleteCustomResources takes a list of unstructured custom resources and
+// sends a request to each one to be deleted.
+func (c *Command) deleteCustomResources(crs []unstructured.Unstructured, kindToResource map[string]string, uiLogger action.DebugLog) error {
+	for _, cr := range crs {
+		gv, err := schema.ParseGroupVersion(cr.GetAPIVersion())
+		if err != nil {
+			return err
+		}
+
+		target := schema.GroupVersionResource{
+			Group:    gv.Group,
+			Version:  gv.Version,
+			Resource: kindToResource[cr.GetKind()],
+		}
+
+		uiLogger(fmt.Sprintf("Starting delete for \"%s\" %s", cr.GetName(), cr.GetKind()))
+		err = c.dynamicK8sClient.
+			Resource(target).
+			Namespace(cr.GetNamespace()).
+			Delete(c.Ctx, cr.GetName(), metav1.DeleteOptions{})
+		if err != nil && !k8serrors.IsNotFound(err) {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// patchCustomResources takes a list of unstructured custom resources and
+// sends a request to each one to patch its finalizers to an empty list.
+func (c *Command) patchCustomResources(crs []unstructured.Unstructured, kindToResource map[string]string, uiLogger action.DebugLog) error {
+	finalizerPatch := []byte(`[{
+		"op": "replace",
+		"path": "/metadata/finalizers",
+		"value": []
+	}]`)
+
+	for _, cr := range crs {
+		gv, err := schema.ParseGroupVersion(cr.GetAPIVersion())
+		if err != nil {
+			return err
+		}
+
+		target := schema.GroupVersionResource{
+			Group:    gv.Group,
+			Version:  gv.Version,
+			Resource: kindToResource[cr.GetKind()],
+		}
+
+		uiLogger(fmt.Sprintf("Patching finalizers for \"%s\" %s", cr.GetName(), cr.GetKind()))
+		_, err = c.dynamicK8sClient.
+			Resource(target).
+			Namespace(cr.GetNamespace()).
+			Patch(c.Ctx, cr.GetName(), types.JSONPatchType, finalizerPatch, metav1.PatchOptions{})
+		if err != nil && !k8serrors.IsNotFound(err) {
+			return err
+		}
+	}
+
 	return nil
 }
 
@@ -414,7 +603,7 @@ func (c *Command) findExistingInstallation(options *helm.CheckForInstallationsOp
 func (c *Command) deletePVCs(foundReleaseName, foundReleaseNamespace string) error {
 	var pvcNames []string
 	pvcSelector := metav1.ListOptions{LabelSelector: fmt.Sprintf("release=%s", foundReleaseName)}
-	pvcs, err := c.kubernetes.CoreV1().PersistentVolumeClaims(foundReleaseNamespace).List(c.Ctx, pvcSelector)
+	pvcs, err := c.k8sClient.CoreV1().PersistentVolumeClaims(foundReleaseNamespace).List(c.Ctx, pvcSelector)
 	if err != nil {
 		return fmt.Errorf("deletePVCs: %s", err)
 	}
@@ -423,14 +612,14 @@ func (c *Command) deletePVCs(foundReleaseName, foundReleaseNamespace string) err
 		return nil
 	}
 	for _, pvc := range pvcs.Items {
-		err := c.kubernetes.CoreV1().PersistentVolumeClaims(foundReleaseNamespace).Delete(c.Ctx, pvc.Name, metav1.DeleteOptions{})
+		err := c.k8sClient.CoreV1().PersistentVolumeClaims(foundReleaseNamespace).Delete(c.Ctx, pvc.Name, metav1.DeleteOptions{})
 		if err != nil {
 			return fmt.Errorf("deletePVCs: error deleting PVC %q: %s", pvc.Name, err)
 		}
 		pvcNames = append(pvcNames, pvc.Name)
 	}
 	err = backoff.Retry(func() error {
-		pvcs, err := c.kubernetes.CoreV1().PersistentVolumeClaims(foundReleaseNamespace).List(c.Ctx, pvcSelector)
+		pvcs, err := c.k8sClient.CoreV1().PersistentVolumeClaims(foundReleaseNamespace).List(c.Ctx, pvcSelector)
 		if err != nil {
 			return fmt.Errorf("deletePVCs: %s", err)
 		}
@@ -453,7 +642,7 @@ func (c *Command) deletePVCs(foundReleaseName, foundReleaseNamespace string) err
 
 // deleteSecrets deletes any secrets that have the label "managed-by" set to "consul-k8s".
 func (c *Command) deleteSecrets(foundReleaseNamespace string) error {
-	secrets, err := c.kubernetes.CoreV1().Secrets(foundReleaseNamespace).List(c.Ctx, metav1.ListOptions{
+	secrets, err := c.k8sClient.CoreV1().Secrets(foundReleaseNamespace).List(c.Ctx, metav1.ListOptions{
 		LabelSelector: common.CLILabelKey + "=" + common.CLILabelValue,
 	})
 	if err != nil {
@@ -465,7 +654,7 @@ func (c *Command) deleteSecrets(foundReleaseNamespace string) error {
 	}
 	var secretNames []string
 	for _, secret := range secrets.Items {
-		err := c.kubernetes.CoreV1().Secrets(foundReleaseNamespace).Delete(c.Ctx, secret.Name, metav1.DeleteOptions{})
+		err := c.k8sClient.CoreV1().Secrets(foundReleaseNamespace).Delete(c.Ctx, secret.Name, metav1.DeleteOptions{})
 		if err != nil {
 			return fmt.Errorf("deleteSecrets: error deleting Secret %q: %s", secret.Name, err)
 		}
@@ -484,7 +673,7 @@ func (c *Command) deleteSecrets(foundReleaseNamespace string) error {
 func (c *Command) deleteServiceAccounts(foundReleaseName, foundReleaseNamespace string) error {
 	var serviceAccountNames []string
 	saSelector := metav1.ListOptions{LabelSelector: fmt.Sprintf("release=%s", foundReleaseName)}
-	sas, err := c.kubernetes.CoreV1().ServiceAccounts(foundReleaseNamespace).List(c.Ctx, saSelector)
+	sas, err := c.k8sClient.CoreV1().ServiceAccounts(foundReleaseNamespace).List(c.Ctx, saSelector)
 	if err != nil {
 		return fmt.Errorf("deleteServiceAccounts: %s", err)
 	}
@@ -493,7 +682,7 @@ func (c *Command) deleteServiceAccounts(foundReleaseName, foundReleaseNamespace 
 		return nil
 	}
 	for _, sa := range sas.Items {
-		err := c.kubernetes.CoreV1().ServiceAccounts(foundReleaseNamespace).Delete(c.Ctx, sa.Name, metav1.DeleteOptions{})
+		err := c.k8sClient.CoreV1().ServiceAccounts(foundReleaseNamespace).Delete(c.Ctx, sa.Name, metav1.DeleteOptions{})
 		if err != nil {
 			return fmt.Errorf("deleteServiceAccounts: error deleting ServiceAccount %q: %s", sa.Name, err)
 		}
@@ -512,7 +701,7 @@ func (c *Command) deleteServiceAccounts(foundReleaseName, foundReleaseNamespace 
 func (c *Command) deleteRoles(foundReleaseName, foundReleaseNamespace string) error {
 	var roleNames []string
 	roleSelector := metav1.ListOptions{LabelSelector: fmt.Sprintf("release=%s", foundReleaseName)}
-	roles, err := c.kubernetes.RbacV1().Roles(foundReleaseNamespace).List(c.Ctx, roleSelector)
+	roles, err := c.k8sClient.RbacV1().Roles(foundReleaseNamespace).List(c.Ctx, roleSelector)
 	if err != nil {
 		return fmt.Errorf("deleteRoles: %s", err)
 	}
@@ -521,7 +710,7 @@ func (c *Command) deleteRoles(foundReleaseName, foundReleaseNamespace string) er
 		return nil
 	}
 	for _, role := range roles.Items {
-		err := c.kubernetes.RbacV1().Roles(foundReleaseNamespace).Delete(c.Ctx, role.Name, metav1.DeleteOptions{})
+		err := c.k8sClient.RbacV1().Roles(foundReleaseNamespace).Delete(c.Ctx, role.Name, metav1.DeleteOptions{})
 		if err != nil {
 			return fmt.Errorf("deleteRoles: error deleting Role %q: %s", role.Name, err)
 		}
@@ -540,7 +729,7 @@ func (c *Command) deleteRoles(foundReleaseName, foundReleaseNamespace string) er
 func (c *Command) deleteRoleBindings(foundReleaseName, foundReleaseNamespace string) error {
 	var rolebindingNames []string
 	rolebindingSelector := metav1.ListOptions{LabelSelector: fmt.Sprintf("release=%s", foundReleaseName)}
-	rolebindings, err := c.kubernetes.RbacV1().RoleBindings(foundReleaseNamespace).List(c.Ctx, rolebindingSelector)
+	rolebindings, err := c.k8sClient.RbacV1().RoleBindings(foundReleaseNamespace).List(c.Ctx, rolebindingSelector)
 	if err != nil {
 		return fmt.Errorf("deleteRoleBindings: %s", err)
 	}
@@ -549,7 +738,7 @@ func (c *Command) deleteRoleBindings(foundReleaseName, foundReleaseNamespace str
 		return nil
 	}
 	for _, rolebinding := range rolebindings.Items {
-		err := c.kubernetes.RbacV1().RoleBindings(foundReleaseNamespace).Delete(c.Ctx, rolebinding.Name, metav1.DeleteOptions{})
+		err := c.k8sClient.RbacV1().RoleBindings(foundReleaseNamespace).Delete(c.Ctx, rolebinding.Name, metav1.DeleteOptions{})
 		if err != nil {
 			return fmt.Errorf("deleteRoleBindings: error deleting Role %q: %s", rolebinding.Name, err)
 		}
@@ -568,7 +757,7 @@ func (c *Command) deleteRoleBindings(foundReleaseName, foundReleaseNamespace str
 func (c *Command) deleteJobs(foundReleaseName, foundReleaseNamespace string) error {
 	var jobNames []string
 	jobSelector := metav1.ListOptions{LabelSelector: fmt.Sprintf("release=%s", foundReleaseName)}
-	jobs, err := c.kubernetes.BatchV1().Jobs(foundReleaseNamespace).List(c.Ctx, jobSelector)
+	jobs, err := c.k8sClient.BatchV1().Jobs(foundReleaseNamespace).List(c.Ctx, jobSelector)
 	if err != nil {
 		return fmt.Errorf("deleteJobs: %s", err)
 	}
@@ -577,7 +766,7 @@ func (c *Command) deleteJobs(foundReleaseName, foundReleaseNamespace string) err
 		return nil
 	}
 	for _, job := range jobs.Items {
-		err := c.kubernetes.BatchV1().Jobs(foundReleaseNamespace).Delete(c.Ctx, job.Name, metav1.DeleteOptions{})
+		err := c.k8sClient.BatchV1().Jobs(foundReleaseNamespace).Delete(c.Ctx, job.Name, metav1.DeleteOptions{})
 		if err != nil {
 			return fmt.Errorf("deleteJobs: error deleting Job %q: %s", job.Name, err)
 		}
@@ -596,7 +785,7 @@ func (c *Command) deleteJobs(foundReleaseName, foundReleaseNamespace string) err
 func (c *Command) deleteClusterRoles(foundReleaseName string) error {
 	var clusterRolesNames []string
 	clusterRolesSelector := metav1.ListOptions{LabelSelector: fmt.Sprintf("release=%s", foundReleaseName)}
-	clusterRoles, err := c.kubernetes.RbacV1().ClusterRoles().List(c.Ctx, clusterRolesSelector)
+	clusterRoles, err := c.k8sClient.RbacV1().ClusterRoles().List(c.Ctx, clusterRolesSelector)
 	if err != nil {
 		return fmt.Errorf("deleteClusterRoles: %s", err)
 	}
@@ -605,7 +794,7 @@ func (c *Command) deleteClusterRoles(foundReleaseName string) error {
 		return nil
 	}
 	for _, clusterRole := range clusterRoles.Items {
-		err := c.kubernetes.RbacV1().ClusterRoles().Delete(c.Ctx, clusterRole.Name, metav1.DeleteOptions{})
+		err := c.k8sClient.RbacV1().ClusterRoles().Delete(c.Ctx, clusterRole.Name, metav1.DeleteOptions{})
 		if err != nil {
 			return fmt.Errorf("deleteClusterRoles: error deleting cluster role %q: %s", clusterRole.Name, err)
 		}
@@ -624,7 +813,7 @@ func (c *Command) deleteClusterRoles(foundReleaseName string) error {
 func (c *Command) deleteClusterRoleBindings(foundReleaseName string) error {
 	var clusterRoleBindingsNames []string
 	clusterRoleBindingsSelector := metav1.ListOptions{LabelSelector: fmt.Sprintf("release=%s", foundReleaseName)}
-	clusterRoleBindings, err := c.kubernetes.RbacV1().ClusterRoleBindings().List(c.Ctx, clusterRoleBindingsSelector)
+	clusterRoleBindings, err := c.k8sClient.RbacV1().ClusterRoleBindings().List(c.Ctx, clusterRoleBindingsSelector)
 	if err != nil {
 		return fmt.Errorf("deleteClusterRoleBindings: %s", err)
 	}
@@ -633,7 +822,7 @@ func (c *Command) deleteClusterRoleBindings(foundReleaseName string) error {
 		return nil
 	}
 	for _, clusterRoleBinding := range clusterRoleBindings.Items {
-		err := c.kubernetes.RbacV1().ClusterRoleBindings().Delete(c.Ctx, clusterRoleBinding.Name, metav1.DeleteOptions{})
+		err := c.k8sClient.RbacV1().ClusterRoleBindings().Delete(c.Ctx, clusterRoleBinding.Name, metav1.DeleteOptions{})
 		if err != nil {
 			return fmt.Errorf("deleteClusterRoleBindings: error deleting cluster role binding %q: %s", clusterRoleBinding.Name, err)
 		}
@@ -646,4 +835,16 @@ func (c *Command) deleteClusterRoleBindings(foundReleaseName string) error {
 		c.UI.Output("Consul cluster role bindings deleted.", terminal.WithSuccessStyle())
 	}
 	return nil
+}
+
+// mapCRKindToResourceName takes the list of custom resource definitions and
+// creates a mapping from the "kind" of the CRD to its "resource" name.
+// This is needed for the dynamic API which finds custom resources by their
+// lowercase, plural resource name. (e.g. "ingressgateways" for "IngressGateway" kind).
+func mapCRKindToResourceName(crds *apiextv1.CustomResourceDefinitionList) map[string]string {
+	kindToResourceName := make(map[string]string)
+	for _, crd := range crds.Items {
+		kindToResourceName[crd.Spec.Names.Kind] = crd.Spec.Names.Plural
+	}
+	return kindToResourceName
 }